move h595 functions to lib
